+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/ParameterSpec.kt
@@ -54,10 +54,8 @@ public class ParameterSpec private constructor(
   internal fun emit(
     codeWriter: CodeWriter,
     includeType: Boolean = true,
-    emitKdoc: Boolean = false,
     inlineAnnotations: Boolean = true,
   ) {
-    if (emitKdoc) codeWriter.emitKdoc(kdoc.ensureEndsWithNewLine())
     codeWriter.emitAnnotations(annotations, inlineAnnotations)
     codeWriter.emitModifiers(modifiers)
     if (name.isNotEmpty()) codeWriter.emitCode("%N", this)