+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/TypeSpec.kt
@@ -48,7 +48,8 @@ public class TypeSpec private constructor(
   ContextReceivable by contextReceivers,
   Annotatable,
   Documentable,
-  TypeSpecHolder {
+  TypeSpecHolder,
+  MemberSpecHolder {
   public val kind: Kind = builder.kind
   public val name: String? = builder.name
   override val kdoc: CodeBlock = builder.kdoc.build()
@@ -73,10 +74,10 @@ public class TypeSpec private constructor(
    */
   public val superinterfaces: Map<TypeName, CodeBlock?> = builder.superinterfaces.toImmutableMap()
   public val enumConstants: Map<String, TypeSpec> = builder.enumConstants.toImmutableMap()
-  public val propertySpecs: List<PropertySpec> = builder.propertySpecs.toImmutableList()
+  override val propertySpecs: List<PropertySpec> = builder.propertySpecs.toImmutableList()
   public val initializerBlock: CodeBlock = builder.initializerBlock.build()
   public val initializerIndex: Int = builder.initializerIndex
-  public val funSpecs: List<FunSpec> = builder.funSpecs.toImmutableList()
+  override val funSpecs: List<FunSpec> = builder.funSpecs.toImmutableList()
   public override val typeSpecs: List<TypeSpec> = builder.typeSpecs.toImmutableList()
   internal val nestedTypesSimpleNames = typeSpecs.map { it.name }.toImmutableSet()
 
@@ -178,6 +179,7 @@ public class TypeSpec private constructor(
         }
         codeWriter.emitTypeVariables(typeVariables)
 
+        var wrapSupertypes = false
         primaryConstructor?.let {
           codeWriter.pushType(this) // avoid name collisions when emitting primary constructor
           val emittedAnnotations = it.annotations.isNotEmpty()
@@ -198,6 +200,8 @@ public class TypeSpec private constructor(
           }
 
           it.parameters.emit(codeWriter, forceNewLines = true) { param ->
+            wrapSupertypes = true
+
             val property = constructorProperties[param.name]
             if (property != null) {
               property.emit(
@@ -209,7 +213,7 @@ public class TypeSpec private constructor(
               )
               param.emitDefaultValue(codeWriter)
             } else {
-              param.emit(codeWriter, emitKdoc = true, inlineAnnotations = false)
+              param.emit(codeWriter, inlineAnnotations = false)
             }
           }
 
@@ -232,7 +236,8 @@ public class TypeSpec private constructor(
         }
 
         if (superTypes.isNotEmpty()) {
-          codeWriter.emitCode(superTypes.joinToCode(separator = ", ", prefix = " : "))
+          val separator = if (wrapSupertypes) ",\n    " else ", "
+          codeWriter.emitCode(superTypes.joinToCode(separator = separator, prefix = " : "))
         }
 
         codeWriter.emitWhereBlock(typeVariables)
@@ -374,10 +379,7 @@ public class TypeSpec private constructor(
   /**
    * Returns KDoc comments including those of the primary constructor and its parameters.
    *
-   * If the primary constructor parameter is not mapped to a property, or if the property doesn't
-   * have its own KDoc - the parameter's KDoc will be attached to the parameter. Otherwise, if both
-   * the parameter and the property have KDoc - the property's KDoc will be attached to the
-   * property/parameter, and the parameter's KDoc will be printed in the type header.
+   * Parameters' KDocs, if present, will always be printed in the type header.
    */
   private fun kdocWithConstructorDocs(): CodeBlock {
     val classKdoc = kdoc.ensureEndsWithNewLine()
@@ -386,10 +388,8 @@ public class TypeSpec private constructor(
         if (primaryConstructor.kdoc.isNotEmpty()) {
           add("@constructor %L", primaryConstructor.kdoc.ensureEndsWithNewLine())
         }
-        val constructorProperties = constructorProperties()
         primaryConstructor.parameters.forEach { parameter ->
-          val propertyKdoc = constructorProperties[parameter.name]?.kdoc ?: CodeBlock.EMPTY
-          if (parameter.kdoc.isNotEmpty() && propertyKdoc.isNotEmpty() && parameter.kdoc != propertyKdoc) {
+          if (parameter.kdoc.isNotEmpty()) {
             add("@param %L %L", parameter.name, parameter.kdoc.ensureEndsWithNewLine())
           }
         }
@@ -476,7 +476,8 @@ public class TypeSpec private constructor(
     ContextReceivable.Builder<Builder>,
     Annotatable.Builder<Builder>,
     Documentable.Builder<Builder>,
-    TypeSpecHolder.Builder<Builder> {
+    TypeSpecHolder.Builder<Builder>,
+    MemberSpecHolder.Builder<Builder> {
     internal var primaryConstructor: FunSpec? = null
     internal var superclass: TypeName = ANY
     internal val initializerBlock = CodeBlock.builder()
@@ -541,6 +542,13 @@ public class TypeSpec private constructor(
           check(primaryConstructor.parameters.size == 1) {
             "value/inline classes must have 1 parameter in constructor"
           }
+        }
+
+        require(
+          primaryConstructor.delegateConstructor == null &&
+            primaryConstructor.delegateConstructorArguments.isEmpty(),
+        ) {
+          "primary constructor can't delegate to other constructors"
         }
       }
       this.primaryConstructor = primaryConstructor
@@ -658,11 +666,7 @@ public class TypeSpec private constructor(
       enumConstants[name] = typeSpec
     }
 
-    public fun addProperties(propertySpecs: Iterable<PropertySpec>): Builder = apply {
-      propertySpecs.map(this::addProperty)
-    }
-
-    public fun addProperty(propertySpec: PropertySpec): Builder = apply {
+    override fun addProperty(propertySpec: PropertySpec): Builder = apply {
       if (EXPECT in modifiers) {
         require(propertySpec.initializer == null) {
           "properties in expect classes can't have initializers"
@@ -679,32 +683,6 @@ public class TypeSpec private constructor(
       propertySpecs += propertySpec
     }
 
-    public fun addProperty(name: String, type: TypeName, vararg modifiers: KModifier): Builder =
-      addProperty(PropertySpec.builder(name, type, *modifiers).build())
-
-    @DelicateKotlinPoetApi(
-      message = "Java reflection APIs don't give complete information on Kotlin types. Consider " +
-        "using the kotlinpoet-metadata APIs instead.",
-    )
-    public fun addProperty(name: String, type: Type, vararg modifiers: KModifier): Builder =
-      addProperty(name, type.asTypeName(), *modifiers)
-
-    public fun addProperty(name: String, type: KClass<*>, vararg modifiers: KModifier): Builder =
-      addProperty(name, type.asTypeName(), *modifiers)
-
-    public fun addProperty(name: String, type: TypeName, modifiers: Iterable<KModifier>): Builder =
-      addProperty(PropertySpec.builder(name, type, modifiers).build())
-
-    @DelicateKotlinPoetApi(
-      message = "Java reflection APIs don't give complete information on Kotlin types. Consider " +
-        "using the kotlinpoet-metadata APIs instead.",
-    )
-    public fun addProperty(name: String, type: Type, modifiers: Iterable<KModifier>): Builder =
-      addProperty(name, type.asTypeName(), modifiers)
-
-    public fun addProperty(name: String, type: KClass<*>, modifiers: Iterable<KModifier>): Builder =
-      addProperty(name, type.asTypeName(), modifiers)
-
     public fun addInitializerBlock(block: CodeBlock): Builder = apply {
       checkCanHaveInitializerBlocks()
       // Set index to however many properties we have
@@ -718,11 +696,7 @@ public class TypeSpec private constructor(
         .add("}\n")
     }
 
-    public fun addFunctions(funSpecs: Iterable<FunSpec>): Builder = apply {
-      funSpecs.forEach { addFunction(it) }
-    }
-
-    public fun addFunction(funSpec: FunSpec): Builder = apply {
+    override fun addFunction(funSpec: FunSpec): Builder = apply {
       funSpecs += funSpec
     }
 
@@ -764,6 +738,43 @@ public class TypeSpec private constructor(
 
     @Suppress("RedundantOverride")
     override fun addTypes(typeSpecs: Iterable<TypeSpec>): Builder = super.addTypes(typeSpecs)
+
+    @Suppress("RedundantOverride")
+    override fun addProperties(propertySpecs: Iterable<PropertySpec>): Builder =
+      super.addProperties(propertySpecs)
+
+    @Suppress("RedundantOverride")
+    override fun addProperty(name: String, type: TypeName, vararg modifiers: KModifier): Builder =
+      super.addProperty(name, type, *modifiers)
+
+    @DelicateKotlinPoetApi(
+      message = "Java reflection APIs don't give complete information on Kotlin types. Consider " +
+        "using the kotlinpoet-metadata APIs instead.",
+    )
+    override fun addProperty(name: String, type: Type, vararg modifiers: KModifier): Builder =
+      super.addProperty(name, type, *modifiers)
+
+    @Suppress("RedundantOverride")
+    override fun addProperty(name: String, type: KClass<*>, vararg modifiers: KModifier): Builder =
+      super.addProperty(name, type, *modifiers)
+
+    @Suppress("RedundantOverride")
+    override fun addProperty(name: String, type: TypeName, modifiers: Iterable<KModifier>): Builder =
+      super.addProperty(name, type, modifiers)
+
+    @DelicateKotlinPoetApi(
+      message = "Java reflection APIs don't give complete information on Kotlin types. Consider " +
+        "using the kotlinpoet-metadata APIs instead.",
+    )
+    override fun addProperty(name: String, type: Type, modifiers: Iterable<KModifier>): Builder =
+      super.addProperty(name, type, modifiers)
+
+    @Suppress("RedundantOverride")
+    override fun addProperty(name: String, type: KClass<*>, modifiers: Iterable<KModifier>): Builder =
+      super.addProperty(name, type, modifiers)
+
+    @Suppress("RedundantOverride")
+    override fun addFunctions(funSpecs: Iterable<FunSpec>): Builder = super.addFunctions(funSpecs)
     //endregion
 
     public fun build(): TypeSpec {